Fixed check for null images in slideImageWithPercentage method
The current slideImageWithPercentage method has one check to see if the
image is null in the animateWithOffset method, but not in the other
areas where slideImageWithPercentage is called.  This check to see if
the imageName is nil is added to suppress the following error:

CUICatalog: Invalid asset name supplied: (null), or invalid scale
factor: 2.000000

This error is a result of trying to perfrom actions on a UIImage that
has a null value.