Used property getter for shouldDrag instead of ivar for more flexibility
We use the property accessor of shouldDrag instead of the ivar so that user can specify in his/her implementation easily if a cell can be swiped or not. User can override the getter for shouldDrag in his custom cell, then return a value based on current state, e.g. return NO if table view is in edit mode (self.isEditing is YES).